+++ sdk/eventgrid/arm-eventgrid/CHANGELOG.md
@@ -1,15 +1,40 @@
 # Release History
+    
+## 14.0.0 (2022-02-28)
+    
+**Features**
 
-## 13.0.1 (Unreleased)
+  - Added Type Alias IsNotNullAdvancedFilter
+  - Added Type Alias IsNullOrUndefinedAdvancedFilter
+  - Added Type Alias NumberInRangeAdvancedFilter
+  - Added Type Alias NumberNotInRangeAdvancedFilter
+  - Added Type Alias PrivateEndpointConnectionsParentType
+  - Added Type Alias StringNotBeginsWithAdvancedFilter
+  - Added Type Alias StringNotContainsAdvancedFilter
+  - Added Type Alias StringNotEndsWithAdvancedFilter
+  - Added Enum KnownPrivateEndpointConnectionsParentType
+  - Enum KnownAdvancedFilterOperatorType has a new value IsNotNull
+  - Enum KnownAdvancedFilterOperatorType has a new value IsNullOrUndefined
+  - Enum KnownAdvancedFilterOperatorType has a new value NumberInRange
+  - Enum KnownAdvancedFilterOperatorType has a new value NumberNotInRange
+  - Enum KnownAdvancedFilterOperatorType has a new value StringNotBeginsWith
+  - Enum KnownAdvancedFilterOperatorType has a new value StringNotContains
+  - Enum KnownAdvancedFilterOperatorType has a new value StringNotEndsWith
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Operation PrivateEndpointConnections.beginDelete has a new signature
+  - Operation PrivateEndpointConnections.beginDeleteAndWait has a new signature
+  - Operation PrivateEndpointConnections.beginUpdate has a new signature
+  - Operation PrivateEndpointConnections.beginUpdateAndWait has a new signature
+  - Operation PrivateEndpointConnections.get has a new signature
+  - Operation PrivateEndpointConnections.listByResource has a new signature
+  - Removed Enum KnownEnum18
+  - Removed Enum KnownEnum19
+  - Removed Enum KnownEnum20
+  - Removed Enum KnownEnum21
+    
+    
 ## 13.0.0 (2021-12-09)
 
 The package of @azure/arm-eventgrid is using our next generation design principles since version 13.0.0, which contains breaking changes.